assets/fastapi-template/requirements.txt
fastapi[all]
python-dotenv
